+++ packages/blockchain-wallet-v4/src/redux/kvStore/lockbox/selectors.js
@@ -99,10 +99,7 @@ export const getLockboxBchContext = state => {
 }
 
 export const getLockboxBchXpub = (state, deviceIndex) =>
-  getDevice(state, deviceIndex)
-    .map(path(['bch', 'accounts']))
-    .map(head)
-    .map(prop('xpub'))
+  getDevice(state, deviceIndex).map(path(['bch', 'accounts', 0, 'xpub']))
 
 export const getBchContextForDevice = (state, deviceIndex) =>
   getDevice(state, deviceIndex)
